Fighters & Watchers
Created fighters and watchers lists.
Fighters will be a new list of the characters fighting the monsters,
watchers will be a new list of those not fighting the monsters, in
appropriate turn order.  Cycling through the watchers list with a call
to cycling through the fighters list at the end of a watcher's action
(iff a watcher acts) shouldn't be too hard.