Hidden char_stats_frame to start, xp up fix
Hid the char stats frame at the beginning of the game, (had to create a
dummy character to do so), and eliminated initial creation of Dave,
Victoria, and Frank.
Modified character's xp up method to allow xp increase beyond the scope
of a level to overflow into the next (ditto if you have enough xp to
complete that next level, etc).