Output text from Label to Text
Output text widget now a text widget, which wraps at word-length instead
of not at all.  When not appending new text to the end, its insert and
delete methods are inactive so users can't modify the output text
however they want once it's in the widget.  Output now actively displays
the last however many lines, instead of the first, so later commands
don't disappear off the bottom.
Added clear_output methods in options and gui, currently accessed by the
player typing "clear output" somewhere in a command from the Entry
widget, but in the future to be used more often (if not entirely)
internally to the program.