Item Use / Useable
Modified "use" option to parse the remaining words in a way that it
separates out the player and item names (it assumes that only items in a
player's backpack can be used and searches from those items' names).
The gamestate, player, and any remaining words are passed to the item's
"useable" function to see if its use is valid at this moment, and if so
calls the item's "use" function (all items or their parent item-type
class should have these functions).