GUI using PIL fix?
I tried using the PIL module instead of Tkinter's PhotoImage to render
the background image, and it seems to work.  PIL does have a really nice
resize method that means we don't have to mess around with zooming to
many times the user's screen size and then subsampling til we get back
down to it.  It also seems pretty fast, though you may have to download
PIL if you don't have it already.