End Turn & item_initialize fix
Changed "Set Action" to "End Turn"; when the first player is added it is
their turn, and when they say "end turn" the turn passes to the next
player in line, initially showing that next player in the char stats
frame.
Fixed gameplay's item_initialize to correctly handle bad item-name-text;
instead of creating an error, it displays a helpful message and discards
the text.