Show_last and Race/Class combos
Implemented the show_last method in options, which references the
Options superclass's last_shown attribute.  This is explicitly set when
the player asks to show something in the options widget, and can be
coded in at the end of any particular option, as in Loot. Options widget
gets cleared at the beginning of each turn. These changes were made per
discussion on 8/18.
Combined "races" and "classes" modules into one, "races_and_classes", to
eliminate circular import statements.
Implemented class and race approved combinations, with an error message
that propagates up from a misused item (which doesn't get consumed) to
the top-level Use option's use, printing out to the gui's output.