Equipment start and xp underflow error-fix
Starting to setup methods for equipping and unequipping equipment, as
well as first, base equipment of all types.  Not yet implemented as
available options, so it shouldn't cause any problems yet.
Fixed xp error for adding xp using bones and not getting the overflowing
xp as xp toward the next level.
Also fixed text-parsing system so that it checks for they options'
keywords individually and as a string of text.  Now options that have
overlapping keywords, such as "equip" and "unequip", shouldn't interfere
in use.